fix(showHide): More robust resize implementation.
The existing resize implementation for the ngShow/ngHide directives
which was used by the mdVirtualRepeat directive incorrectly fired
the resize code when ANY ngShow/ngHide changed on the scope or any
child scope.

Implement a more robust resize solution which checks for the
existence of listeners on that scope before firing and notifying
children.

Fixes #7159.